Tehty 4.11

+++ osa4/blogilista/controllers/blogs.js
@@ -5,6 +5,8 @@ const mongoose = require('mongoose')
 require("dotenv/config")
 const express = require("express")
 const blogsRouter = express.Router()
+const User = require("../models/user")
+const jwt = require('jsonwebtoken')
 
 const mongoUrl = config.MONGODB_URI
 mongoose.connect(mongoUrl)
@@ -13,22 +15,59 @@ blogsRouter.use(cors())
 blogsRouter.use(express.json())
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate("user", { username: 1, name: 1, id: 1 })
+
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  if (!blog.title && !blog.url) {
-    response.status(400).send({ error: "missing fields"})
-  } else {
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
+
+  if (!request.body.title && !request.body.url) {
+    return response.status(400).json({ error: "missing fields" })
   }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = request.user
+  
+  const blog = new Blog({
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes ? request.body.likes : 0,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
   const id = request.params.id
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = request.user
+  const blog = await Blog.findById(id)
+  console.log("user: " + user)
+  console.log("blog: " + blog)
+
+  if (!blog) {
+    return response.status(404).json({ error: "blog not found" })
+  } else if (user.id.toString() !== blog.user.toString()) {
+    return response.status(400).json({ error: 'can\'t delete someone else\'s blog' })
+  }
+
   await Blog.findByIdAndDelete(id)
   response.status(204).end()
 })
